+++ CHANGELOG.md
@@ -1,6 +1,22 @@
 # Hanami::Helpers
 View helpers for Ruby web applications
 
+## v1.1.1 - 2017-11-28
+### Fixed
+- [Alfonso Uceda] Ensure `#select` form helper to not select options with `nil` value
+- [Alfonso Uceda] Ensure `#fields_for_collection` form helper to produce input fields with correct `name` attribute
+- [Luca Guidi] Ensure `#select` form helper to respect `:selected` option
+
+## v1.1.0 - 2017-10-25
+
+## v1.1.0.rc1 - 2017-10-16
+
+## v1.1.0.beta3 - 2017-10-04
+
+## v1.1.0.beta2 - 2017-10-03
+
+## v1.1.0.beta1 - 2017-08-11
+
 ## v1.0.0 - 2017-04-06
 
 ## v1.0.0.rc1 - 2017-03-31